Modularize file utils
