+++ dreamos/core/utils/exceptions.py
@@ -1,0 +1,5 @@
+"""Custom exceptions for file operations."""
+
+
+class FileOperationError(Exception):
+    """Raised when a file operation fails."""
