+++ dreamos/core/utils/file_utils.py
@@ -1,552 +1,59 @@
-"""
-File Utilities
-------------
-Core file operations for Dream.OS system.
-"""
+"""Aggregate helpers for backward compatibility."""
 
-import json
-import logging
-import os
-import shutil
-import aiofiles
-import asyncio
-import yaml
-from contextlib import contextmanager
-from pathlib import Path
-from typing import Any, Dict, Optional, Union, List
-from datetime import datetime
+from .safe_io import (
+    atomic_write,
+    safe_read,
+    safe_write,
+    async_atomic_write,
+    async_delete_file,
+)
 
-logger = logging.getLogger(__name__)
+from .serialization import (
+    load_json,
+    save_json,
+    read_json,
+    write_json,
+    restore_backup,
+    read_yaml,
+    load_yaml,
+    write_yaml,
+    save_yaml,
+    async_save_json,
+    async_load_json,
+)
 
-def atomic_write(file_path: Union[str, Path], content: str, mode: str = 'w') -> bool:
-    """Write content to file atomically.
-    
-    Args:
-        file_path: Path to write to
-        content: Content to write
-        mode: File open mode
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        temp_path = f"{file_path}.tmp"
-        with open(temp_path, mode) as f:
-            f.write(content)
-        os.replace(temp_path, file_path)
-        return True
-    except Exception as e:
-        logger.error(f"Error in atomic write to {file_path}: {e}")
-        return False
+from .file_ops import (
+    ensure_dir,
+    clear_dir,
+    archive_file,
+    extract_agent_id,
+    backup_file,
+    safe_rmdir,
+    cleanup_old_files,
+)
 
-def safe_read(file_path: Union[str, Path], default: Any = None) -> Any:
-    """Safely read content from file.
-    
-    Args:
-        file_path: Path to read from
-        default: Default value if read fails
-        
-    Returns:
-        File content or default value
-    """
-    try:
-        with open(file_path, 'r') as f:
-            return f.read()
-    except Exception as e:
-        logger.error(f"Error reading {file_path}: {e}")
-        return default
-
-def safe_write(file_path: Union[str, Path], content: str) -> bool:
-    """Safely write content to file.
-    
-    Args:
-        file_path: Path to write to
-        content: Content to write
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        with open(file_path, 'w') as f:
-            f.write(content)
-        return True
-    except Exception as e:
-        logger.error(f"Error writing to {file_path}: {e}")
-        return False
-
-def load_json(file_path: Union[str, Path], default: Any = None) -> Any:
-    """Load JSON data from file.
-    
-    Args:
-        file_path: Path to JSON file
-        default: Default value if load fails
-        
-    Returns:
-        Loaded JSON data or default value
-    """
-    try:
-        with open(file_path, 'r') as f:
-            return json.load(f)
-    except Exception as e:
-        logger.error(f"Error loading JSON from {file_path}: {e}")
-        return default
-
-def save_json(file_path: Union[str, Path], data: Any) -> bool:
-    """Save data to JSON file.
-    
-    Args:
-        file_path: Path to save JSON file
-        data: Data to save
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        with open(file_path, 'w') as f:
-            json.dump(data, f, indent=2)
-        return True
-    except Exception as e:
-        logger.error(f"Error saving JSON to {file_path}: {e}")
-        return False
-
-def ensure_dir(path: Union[str, Path]) -> bool:
-    """Ensure directory exists.
-    
-    Args:
-        path: Directory path
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        Path(path).mkdir(parents=True, exist_ok=True)
-        return True
-    except Exception as e:
-        logger.error(f"Error creating directory {path}: {e}")
-        return False
-
-def clear_dir(path: Union[str, Path]) -> bool:
-    """Clear directory contents.
-    
-    Args:
-        path: Directory path
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        for item in Path(path).glob("*"):
-            if item.is_file():
-                item.unlink()
-            elif item.is_dir():
-                shutil.rmtree(item)
-        return True
-    except Exception as e:
-        logger.error(f"Error clearing directory {path}: {e}")
-        return False
-
-def archive_file(file_path: Path, archive_dir: Path, logger: Optional[logging.Logger] = None) -> bool:
-    """Archive a file to the specified directory.
-    
-    Args:
-        file_path: Path to the file to archive
-        archive_dir: Directory to archive to
-        logger: Optional logger instance
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        # Create archive directory if it doesn't exist
-        archive_dir.mkdir(parents=True, exist_ok=True)
-        
-        # Move to archive
-        archive_path = archive_dir / file_path.name
-        file_path.rename(archive_path)
-        
-        # Log success
-        if logger:
-            logger.info(
-                platform="file_ops",
-                status="archived",
-                message=f"Archived file {file_path.name}",
-                tags=["archive", "success"]
-            )
-        return True
-        
-    except Exception as e:
-        # Log error
-        if logger:
-            logger.error(
-                platform="file_ops",
-                status="error",
-                message=f"Error archiving file {file_path}: {e}",
-                tags=["archive", "error"]
-            )
-        return False
-
-def extract_agent_id(file_path: Path) -> Optional[str]:
-    """Extract agent ID from a response file.
-    
-    Args:
-        file_path: Path to the response file
-        
-    Returns:
-        Agent ID if found, None otherwise
-    """
-    try:
-        with open(file_path, 'r') as f:
-            response = json.load(f)
-        return response.get("agent_id")
-    except Exception as e:
-        logger.error(f"Error extracting agent ID from {file_path}: {e}")
-        return None
-
-def read_json(file_path: Union[str, Path], default: Any = None) -> Any:
-    """Read JSON data from file.
-    
-    Args:
-        file_path: Path to JSON file
-        default: Default value if read fails
-        
-    Returns:
-        Loaded JSON data or default value
-    """
-    return load_json(file_path, default)
-
-def backup_file(file_path: Union[str, Path], backup_dir: Union[str, Path], logger: Optional[logging.Logger] = None) -> bool:
-    """Create a backup of a file.
-    
-    Args:
-        file_path: Path to the file to backup
-        backup_dir: Directory to store backup
-        logger: Optional logger instance
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        # Create backup directory if it doesn't exist
-        backup_dir = Path(backup_dir)
-        backup_dir.mkdir(parents=True, exist_ok=True)
-        
-        # Create backup
-        backup_path = backup_dir / f"{Path(file_path).name}.bak"
-        shutil.copy2(file_path, backup_path)
-        
-        # Log success
-        if logger:
-            logger.info(
-                platform="file_ops",
-                status="backed_up",
-                message=f"Backed up file {file_path}",
-                tags=["backup", "success"]
-            )
-        return True
-        
-    except Exception as e:
-        # Log error
-        if logger:
-            logger.error(
-                platform="file_ops",
-                status="error",
-                message=f"Error backing up file {file_path}: {e}",
-                tags=["backup", "error"]
-            )
-        return False
-
-def write_json(data: Any, filepath: str, indent: int = 4) -> bool:
-    """Write data to a JSON file.
-    
-    Args:
-        data: The data to write
-        filepath: Path to the file
-        indent: Number of spaces for indentation
-        
-    Returns:
-        bool: True if successful, False otherwise
-    """
-    try:
-        with open(filepath, 'w', encoding='utf-8') as f:
-            json.dump(data, f, indent=indent)
-        return True
-    except Exception as e:
-        logger.error(f"Failed to write JSON to {filepath}: {e}")
-        return False
-
-def restore_backup(backup_path: str, target_path: str) -> bool:
-    """Restore a file from its backup.
-    
-    Args:
-        backup_path: Path to the backup file
-        target_path: Path where the file should be restored
-        
-    Returns:
-        bool: True if successful, False otherwise
-    """
-    try:
-        if not os.path.exists(backup_path):
-            logger.error(f"Backup file not found: {backup_path}")
-            return False
-            
-        shutil.copy2(backup_path, target_path)
-        logger.info(f"Restored backup from {backup_path} to {target_path}")
-        return True
-    except Exception as e:
-        logger.error(f"Failed to restore backup from {backup_path}: {e}")
-        return False
-
-def read_yaml(file_path: Union[str, Path], default: Any = None) -> Any:
-    """Read YAML data from file.
-    
-    Args:
-        file_path: Path to YAML file
-        default: Default value if read fails
-        
-    Returns:
-        Loaded YAML data or default value
-    """
-    try:
-        with open(file_path, 'r') as f:
-            return yaml.safe_load(f)
-    except Exception as e:
-        logger.error(f"Error reading YAML from {file_path}: {e}")
-        return default
-
-def load_yaml(file_path: Union[str, Path], default: Any = None) -> Any:
-    """Alias for read_yaml for backward compatibility."""
-    return read_yaml(file_path, default)
-
-def write_yaml(file_path: Union[str, Path], data: Any) -> bool:
-    """Write data to YAML file.
-    
-    Args:
-        file_path: Path to write YAML file
-        data: Data to write
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        with open(file_path, 'w') as f:
-            yaml.safe_dump(data, f, default_flow_style=False)
-        return True
-    except Exception as e:
-        logger.error(f"Error writing YAML to {file_path}: {e}")
-        return False
-
-def safe_rmdir(path: Union[str, Path]) -> bool:
-    """Safely remove a directory and its contents.
-    
-    Args:
-        path: Directory path to remove
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        if os.path.exists(path):
-            shutil.rmtree(path)
-        return True
-    except Exception as e:
-        logger.error(f"Error removing directory {path}: {e}")
-        return False
-
-def save_yaml(file_path: Union[str, Path], data: Any) -> bool:
-    """Save data to YAML file.
-    
-    Args:
-        file_path: Path to save YAML file
-        data: Data to save
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        with open(file_path, 'w') as f:
-            yaml.safe_dump(data, f, default_flow_style=False)
-        return True
-    except Exception as e:
-        logger.error(f"Error saving YAML to {file_path}: {e}")
-        return False
-
-async def async_atomic_write(file_path: Union[str, Path], content: str, mode: str = 'w') -> bool:
-    """Write content to file atomically using async I/O.
-    
-    Args:
-        file_path: Path to write to
-        content: Content to write
-        mode: File open mode
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        path = Path(file_path)
-        temp_path = path.with_suffix('.tmp')
-        
-        async with aiofiles.open(temp_path, mode) as f:
-            await f.write(content)
-        
-        os.replace(temp_path, path)
-        return True
-    except Exception as e:
-        logger.error(f"Error in async atomic write to {file_path}: {e}")
-        if temp_path.exists():
-            temp_path.unlink()
-        return False
-
-async def async_delete_file(file_path: Union[str, Path], logger: Optional[logging.Logger] = None) -> bool:
-    """Delete a file using async I/O.
-    
-    Args:
-        file_path: Path to delete
-        logger: Optional logger instance
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        path = Path(file_path)
-        if path.exists():
-            path.unlink()
-            
-            if logger:
-                logger.info(
-                    platform="file_ops",
-                    status="deleted",
-                    message=f"Deleted file {path.name}",
-                    tags=["delete", "success"]
-                )
-            return True
-        return False
-    except Exception as e:
-        if logger:
-            logger.error(
-                platform="file_ops",
-                status="error",
-                message=f"Error deleting file {file_path}: {e}",
-                tags=["delete", "error"]
-            )
-        return False
-
-async def cleanup_old_files(
-    directory: Union[str, Path],
-    max_age_hours: int = 24,
-    pattern: str = "*.*",
-    logger: Optional[logging.Logger] = None
-) -> List[Path]:
-    """Clean up old files in a directory.
-    
-    Args:
-        directory: Directory to clean
-        max_age_hours: Maximum age of files in hours
-        pattern: File pattern to match
-        logger: Optional logger instance
-        
-    Returns:
-        List of deleted file paths
-    """
-    try:
-        dir_path = Path(directory)
-        now = datetime.now()
-        deleted_files = []
-        
-        for filepath in dir_path.glob(pattern):
-            # Get file age
-            age = now - datetime.fromtimestamp(filepath.stat().st_mtime)
-            age_hours = age.total_seconds() / 3600
-            
-            # Delete if too old
-            if age_hours > max_age_hours:
-                if await async_delete_file(filepath, logger):
-                    deleted_files.append(filepath)
-        
-        if logger:
-            logger.info(
-                platform="file_ops",
-                status="cleanup",
-                message=f"Cleaned up {len(deleted_files)} old files",
-                tags=["cleanup", "success"]
-            )
-        
-        return deleted_files
-        
-    except Exception as e:
-        if logger:
-            logger.error(
-                platform="file_ops",
-                status="error",
-                message=f"Error cleaning up old files: {e}",
-                tags=["cleanup", "error"]
-            )
-        return []
-
-async def async_save_json(
-    file_path: Union[str, Path],
-    data: Any,
-    logger: Optional[logging.Logger] = None
-) -> bool:
-    """Save data to JSON file using async I/O.
-    
-    Args:
-        file_path: Path to save JSON file
-        data: Data to save
-        logger: Optional logger instance
-        
-    Returns:
-        True if successful, False otherwise
-    """
-    try:
-        content = json.dumps(data, indent=2)
-        success = await async_atomic_write(file_path, content)
-        
-        if success and logger:
-            logger.info(
-                platform="file_ops",
-                status="saved",
-                message=f"Saved JSON to {Path(file_path).name}",
-                tags=["json", "success"]
-            )
-        
-        return success
-        
-    except Exception as e:
-        if logger:
-            logger.error(
-                platform="file_ops",
-                status="error",
-                message=f"Error saving JSON to {file_path}: {e}",
-                tags=["json", "error"]
-            )
-        return False
-
-async def async_load_json(
-    file_path: Union[str, Path],
-    default: Any = None,
-    logger: Optional[logging.Logger] = None
-) -> Any:
-    """Load JSON data from file using async I/O.
-    
-    Args:
-        file_path: Path to JSON file
-        default: Default value if load fails
-        logger: Optional logger instance
-        
-    Returns:
-        Loaded JSON data or default value
-    """
-    try:
-        async with aiofiles.open(file_path, 'r') as f:
-            content = await f.read()
-            return json.loads(content)
-            
-    except Exception as e:
-        if logger:
-            logger.error(
-                platform="file_ops",
-                status="error",
-                message=f"Error loading JSON from {file_path}: {e}",
-                tags=["json", "error"]
-            )
-        return default 
+__all__ = [
+    "atomic_write",
+    "safe_read",
+    "safe_write",
+    "async_atomic_write",
+    "async_delete_file",
+    "load_json",
+    "save_json",
+    "read_json",
+    "write_json",
+    "restore_backup",
+    "read_yaml",
+    "load_yaml",
+    "write_yaml",
+    "save_yaml",
+    "async_save_json",
+    "async_load_json",
+    "ensure_dir",
+    "clear_dir",
+    "archive_file",
+    "extract_agent_id",
+    "backup_file",
+    "safe_rmdir",
+    "cleanup_old_files",
+]
